// src/enums/enums.ts
enum LoopType {
	For,
	ForEach
}

export enum MessageTypes {
	AbjectFailure = "Abject Failure",
	Error = "Error",
	Warning = "Warning"
}

export enum ElementTypes {
	Variable = "variable. Variables are declared with the const, let or var keyword.",
	Function = "function",
	ConsoleObject = "console object",
	TypeOf = "typeof",
	ForLoop = "for loop",
	ForEachLoop = "forEach loop"
}